Restructured Rent-Main and all files that require the module. Renamed Rent-Main to Rent-API.
