+++ Rent/Rent-API/src/main/java/com/rent/api/ApiServer.java
@@ -1,7 +1,8 @@
+package com.rent.api;
+
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
-import org.springframework.boot.autoconfigure.security.SecurityAutoConfiguration;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
@@ -9,9 +10,7 @@ import org.springframework.context.annotation.Configuration;
 import java.util.Arrays;
 
 @Configuration
-@ComponentScan("Configuration," +
-               "Controllers," +
-               "Services," +
+@ComponentScan("com.rent.api," +
                "Utilities," +
                "com.rent.data.configuration," +
                "com.rent.data.dataaccess.api")
