+++ Rent/Rent-API/src/main/java/com/rent/api/configuration/AnnotationConfigDispatcherServletInitializer.java
@@ -1,4 +1,4 @@
-package Configuration;
+package com.rent.api.configuration;
 
 import org.springframework.web.servlet.support.AbstractAnnotationConfigDispatcherServletInitializer;
