+++ Rent/Rent-API/src/main/java/com/rent/api/configuration/GeneralProperties.java
@@ -1,4 +1,4 @@
-package Configuration;
+package com.rent.api.configuration;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
