+++ Rent/Rent-API/src/main/java/com/rent/api/configuration/RootConfiguration.java
@@ -1,4 +1,4 @@
-package Configuration;
+package com.rent.api.configuration;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
