+++ Rent/pom.xml
@@ -17,7 +17,7 @@
     <modules>
         <module>Rent-Util</module>
         <module>Rent-Data</module>
-        <module>Rent-Main</module>
+        <module>Rent-API</module>
         <module>Rent-Oauth</module>
     </modules>
